We need to loosen up a condition on when to reparse the existing TU.
Previously, we did it only for the cases when we detect that file is
edited (e.g. original_filename != contents_filename) AND if timestamp
has changed. This approach falls short in the following use-case and
failes to provide up-to-date TU:
  1. Open some file
  2. Start editing the file
  3. Undo the changes

Basically, undoing the changes is also an editing action but we will
fail to detect it because our (faulty) assumption was that file has been
edited only if (original_filename != contents_filename) which is obviously
wrong. This resulted in parse() function providing not up-to-date TUnit back
to the client code. To really make sure that file has been or has been not
edited, we have to check its timestamp.